Rename transformer to frontend.
