+++ python/freetensor/core/autograd.py
@@ -6,7 +6,7 @@ import freetensor_ffi as ffi
 from freetensor_ffi import GradTapeMode
 from freetensor_ffi import output_intermediates
 
-from .transformer import transform
+from .frontend import transform
 
 
 class Return:
